+++ src-tauri/src/main.rs
@@ -11,6 +11,7 @@ fn main() {
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_http::init())
+        .plugin(tauri_plugin_opener::init())
         .invoke_handler(tauri::generate_handler![
             gpg::gpg_info,
             gpg::gpg_list_keys,